+++ ogma-core/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Revision history for ogma-core
 
+## [1.X.Y] - 2024-12-04
+
+* Replace queueSize with QUEUE_SIZE in FPP file (#186).
+
 ## [1.5.0] - 2024-11-21
 
 * Version bump 1.5.0 (#178).