+++ ogma-core/CHANGELOG.md
@@ -1,10 +1,11 @@
 # Revision history for ogma-core
 
-## [1.X.Y] - 2024-12-23
+## [1.X.Y] - 2024-12-24
 
 * Replace queueSize with QUEUE_SIZE in FPP file (#186).
 * Use template expansion system to generate F' monitoring component (#185).
 * Use template expansion system to generate standalone Copilot monitor (#189).
+* Add repository information to cabal package (#148).
 
 ## [1.5.0] - 2024-11-21