+++ ogma-core/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Revision history for ogma-core
 
-## [1.X.Y] - 2025-01-14
+## [1.X.Y] - 2025-01-19
 
 * Replace queueSize with QUEUE_SIZE in FPP file (#186).
 * Use template expansion system to generate F' monitoring component (#185).
@@ -9,6 +9,7 @@
 * Add version bounds to all dependencies (#119).
 * Add command to transform state diagrams into monitors (#194).
 * Extend standalone command to use external process to parse properties (#197).
+* Enable using user-provided file as format definition spec (#200).
 
 ## [1.5.0] - 2024-11-21